:sparkles: Getting the core base-reading code in place
Some testing go be added here, and generally some testing to be done.
I *think* this is reading correctly but I need to compare what I'm seeing vs
the same data inside a third party genome browser; or at least the same file
tested against one of my other 2bit readers which I will have validated.

But that's not important right this moment; this is about getting the rough
cut of the core reading code in place.