:racehorse: Try and improve performance while handling mask blocks
Still not as good as I'd like, but a slight improvement. Here I just have
the sequence look at the blocks that are known to intersect with it, while
also having a cache of on the sequence class. Individual cases this isn't
going to make a ton of difference, but if the caller makes repeated calls it
should help.