:bug: Admitting defeat on scrollbar issues for now
It seems that there isn't a reliable way to get the width of a widget,
especially if it starts without a vertical scrollbar and then acquires one,
especially given that the addition of the scrollbar comes about from the
initial calculation. There's sort of a Catch-22 thing going on here.

So to make life easier for now... let's always have one on; it's almost
always going to be necessary anyway.