:bug: Fix an off-by-one error in the output
On top of the scrollbars eating two bases, I still had one base going
missing per line. Turns out I was messing up the calculation of the width of
the label. Or, more to the point, I was using its width including the
same (which is correct), then going on to pad the actual number with that
many spaces *and* then add a space.

D'oh!