:bug: Ensure we always see the last line
I... honestly I'm not sure quite why I need to do it this way, this came
from poking and prodding. I need to think about this properly a little later
but, for now, I just want to get it going so I can carry on with what I want
to get up and running.