:bug: Handle being resized
It sort of worked anyway, but depending on if the display was getting wider
or narrower you could end up with too much empty space, or truncated data,
at the end of the display. This change means the length of the virtual
display gets recalculated.

It does have one unintended consequence though: you seem to get a double
refresh because:

  Resize > Update Virtual Size > Resize Because Virtual Resize > Resize

There might be a way to calm that down, but for now this is fine.